+++ src/views/tasks/task-list/task-list.ts
@@ -1,6 +1,6 @@
 import { ChangeDetectionStrategy, Component, EventEmitter, Input, Output } from '@angular/core';
 import { FirebaseListObservable } from 'angularfire2';
-import { ITask } from 'src/core/task';
+import { ITask } from 'src/core/tasks';
 import { TaskItem } from '../task-item/task-item';
 import { TaskListFilterPipe } from './task-list-filter-pipe';
